# api/openapi.yaml
openapi: 3.0.3
info:
  title: Pennywise API
  version: 1.0.0
  description: Self-hosted personal finance dashboard API
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /projections:
    post:
      operationId: computeProjection
      tags: [projections]
      summary: Compute financial projection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectionRequest"
      responses:
        "200":
          description: Projection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  responses:
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
...
    Scenario:
      type: string
      enum:
        - best
        - average
        - worst
...
    ProjectionRequest:
      type: object
      required:
        - years_to_project
      properties:
        monthly_savings_adjustment:
          type: number
          description: Percentage adjustment to current savings rate
        return_rate:
          type: number
          minimum: 0
          maximum: 100
        years_to_project:
          type: integer
          minimum: 1
          maximum: 40
        extra_debt_payment:
          type: number
          minimum: 0
          description: Extra monthly payment toward highest-rate debt
        one_time_events:
          type: array
          items:
            type: object
            required:
              - amount
              - date
              - type
            properties:
              amount:
                type: number
              date:
                type: string
                format: date
              type:
                type: string
                enum:
                  - windfall
                  - expense
    ProjectionResponse:
      type: object
      required:
        - scenarios
      properties:
        scenarios:
          type: array
          items:
            type: object
            required:
              - scenario
              - data_points
            properties:
              scenario:
                $ref: "#/components/schemas/Scenario"
              data_points:
                type: array
                items:
                  type: object
                  required:
                    - date
                    - value
                  properties:
                    date:
                      type: string
                      format: date
                    value:
                      type: number
              millionaire_date:
                type: string
                format: date
              debt_free_date:
                type: string
                format: date
...